Add fix for initial upload of files
Can't compare if it's the first upload

+++ blanc_basic_assets/assets/listeners.py
@@ -3,12 +3,14 @@ from .models import Image, File
 
 
 def asset_file_change(sender, instance, **kwargs):
-    old_obj = sender.objects.get(pk=instance.pk)
+    # Must be saved already
+    if instance.pk is not None:
+        old_obj = sender.objects.get(pk=instance.pk)
 
-    # Delete the old file if the file names don't match
-    if old_obj.file.name != instance.file.name:
-        storage = old_obj.file.storage
-        storage.delete(name=old_obj.file.name)
+        # Delete the old file if the file names don't match
+        if old_obj.file.name != instance.file.name:
+            storage = old_obj.file.storage
+            storage.delete(name=old_obj.file.name)
 
 
 def asset_file_delete(instance, **kwargs):
